@@ -43,9 +43,9 @@
     "moment": "^2.23.0",
     "node-sass": "^4.9.0",
     "prop-types": "^15.6.1",
-    "react": "^16.6.3",
+    "react": "^16.8.5",
     "react-device-detect": "^1.3.4",
-    "react-dom": "^16.6.3",
+    "react-dom": "^16.8.5",
     "react-fontawesome": "^1.6.1",
     "react-ga": "^2.5.3",
     "react-infinite-scroller": "^1.2.2",
